[Resource] Refactored DoctrineTargetEntitiesResolverSpec as it started to fail
