--- src/Sylius/Bundle/ResourceBundle/spec/Fixture/Document/FooInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace spec\Sylius\Bundle\ResourceBundle\Fixture\Document;
-
-/**
- * Foo Interface.
- *
- * @author Ivannis Suárez Jérez <ivannis.suarez@gmail.com>
- */
-interface FooInterface
-{
-}
