--- src/Sylius/Bundle/ResourceBundle/spec/Fixture/Entity/Foo.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace spec\Sylius\Bundle\ResourceBundle\Fixture\Entity;
-
-/**
- * Foo entity.
- *
- * @author Saša Stamenković <umpirsky@gmail.com>
- */
-class Foo
-{
-}
