--- src/Sylius/Bundle/ResourceBundle/spec/Fixture/Entity/TranslatableFoo.php
@@ -1,27 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace spec\Sylius\Bundle\ResourceBundle\Fixture\Entity;
-
-use Sylius\Component\Translation\Model\AbstractTranslatable;
-
-/**
- * Foo translatable entity.
- *
- * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
- */
-class TranslatableFoo extends AbstractTranslatable
-{
-    public static function getTranslationClass()
-    {
-        return  'spec\Sylius\Bundle\ResourceBundle\Fixture\Entity\Foo';
-    }
-}
